requirements.txt
dj-database-url==0.5.0
Django==2.0.7
django-cors-headers==2.4.0
djangorestframework==3.8.2
djangorestframework-jsonapi==2.5.0
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
inflection==0.3.1
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
pytz==2018.5
six==1.11.0
whitenoise==3.3.1
